Shrinking Data by Half
Update remove_margin function in data3d.h

+++ VesselNess/Graph.h
@@ -41,8 +41,8 @@ public:
 	inline const std::vector<Node_Type>& get_nodes( void ) const { return nodes; }
 	inline       std::vector<Node_Type>& get_nodes( void )       { return nodes; }
 	inline Node_Type& get_node(const int& i){ return nodes[i]; }
-	const unsigned int num_edges(void) const { return edges.size(); }
-	const unsigned int num_nodes(void) const { return nodes.size(); }
+	const unsigned int num_edges(void) const { return (unsigned int) edges.size(); }
+	const unsigned int num_nodes(void) const { return (unsigned int) nodes.size(); }
 
 	// get a minimum spaning tree of the current graph
 	void get_min_span_tree( Graph<Edge_Type, Node_Type>& dst ) const;
